+++ src/test/pages/Loans/repay.test.tsx
@@ -15,8 +15,8 @@ import {
 } from '@/test/mocks/@interlay/interbtc-api/parachain/loans';
 
 import { act, render, screen, userEvent, waitFor } from '../../test-utils';
+import { submitForm, withinModalTabPanel } from '../utils/loans';
 import { TABLES } from './constants';
-import { submitForm, withinModalTabPanel } from './utils';
 
 const path = '/lending';
 const tab = 'repay';
